add failing test to solve for Hall response density matrix
Compiles, but crashes due to the preallocated space for the collision matrix
being insufficient. Will solve this by counting the number of nonzeros in
each row before preallocation.

May need to do something similar for Mat_from_sum_fn. In this case an upper
bound for the nonzero structure could also be provided by querying
the nonzero structure of the Bs.


Former-commit-id: 96c9838492413c9e8e9c77cb177e512ffcc2defe