convert Dbar_B to function: requires complex PetscScalar
Dbar_B(<rho>) is represented by function apply_driving_magnetic(), which takes
as input (DH_0/Dk x \hat{B}) and <rho>. DH_0/Dk has off-diagonal elements which
may be complex. Converted PetscScalar to complex to allow DH_0/Dk to be
represented as a Mat. This conversion is also necessary for the future
introduction of the Berry connection, which may have complex elements, as a
Mat.