k-space sampling density factor fixes, yielding correct Rashba result
(1) Fixed k-grid spacing factor in the derivative operator:
1/(distance between k-points along reciprocal lattice direction d) =
number of k-points along d,
not 1/(number of k-points along d).

(2) Fixed missing factor of 1/(total number of k-points) in evaluation of
trace to give expectation value. This factor is required for expectation
value to be independent of the number of k-points.

Together, these fixes give the expected results for the Rashba model
when chemical potential is near the bottom of the band
(in units where hbar = e = lattice constant = 1):

<s_y> from <n_E^(-1)> = (-t_r / pi * U_0^2)
    \approx -0.06 for t_r = 0.2 t_0, U_0 = 1.0 t_0, t_0 = 1.0.

<j_{s,z}^{y}> from <S_E^(0),int> = -<j_{s,z}^{y}> from <S_E^(0),ext>
    = 1/(8 * pi) \approx = 0.04.

Also gives expected result for Rashba model with magnetic moment:

<\sigma_{xy}> from <S_E^(0),int> = -<\sigma_{xy}> from <S_E^(0),ext>

These have correct signs and relationship but the values have not
been checked here - they do not have chemical potential-independent values
in the continuum model regime.