Allow for non-periodic k-space sampling
In some situations, it is useful for a subsection of the Brillouin
zone to be sampled (in particular, this allows for the
representation of continuum models).
This subsection of the Brillouin zone is represented by the new
`k_min` and `k_max` members of `kmBasis`, defining the boundary
(in reciprocal lattice coordinates) of the sampled region.
`km_at()` is modified to return points sampling this region.

When a finite difference formula for the derivative calls for points
outside the boundary, an alternate finite difference must be used
which is fully contained within the boundary. A scheme for
choosing an appropriate alternative formula is added.
To facilitate this, `kmBasis::add()` is modified to return a
`boost::optional`, which takes the value `boost::none` if the
resulting point would fall outside the boundary.