introduce OwnedMat type and use in matrix creation functions
Introduce a type `OwnedMat` which may own a PETSc `Mat` (i.e. it is
responsible for the destruction of the associated `Mat`, if one is
associated). To avoid expensive unintended copies of the underlying
`Mat`, this type is move-only - it has the semantics of a
`std::unique_ptr`. It may be useful to introduce an explicit
`clone()` operation later which makes a copy of the underlying matrix.

This type is introduced to enable the generalization of `DMGraphNode`
to hold any data on the graph, i.e. a `DMGraphNode<T>` which
associates a `T` with each position on the graph. For a `T` which is
a container, this is much more convenient if we can assume that
`T` will clean itself up, so that the destructor of `DMGraphNode<T>`
does not need special behavior for each `T`.