convert collision matrix to shell and generalize delta function
Memory use is dramatically reduced, with improved scaling: no
process-distributed matrices are larger than Nk * Nm * Nm, and all
process-local structures aside from the eigenvector cache are
(Nk * Nm)-length vectors.

(Memory use may be further improved in the future by using OpenMP to
share constant process-local data. This is particularly important for
the eigenvector cache.)

Elements (km) of [K^{dd} <n>]_k^m and (kmm') of [K^{od} <n>]_k^{mm'} are
skipped when they are known to vanish for the given <n>. This provides
substantial improvement of execution time when the size and number of occupied
Fermi surfaces is small (i.e. when beta is large, and when sigma is small).
Execution time is increased in some cases which are more favorable to the
previous approach in which K^{dd} is pre-calculated and re-used for each
chemical potential value.

Refinement of the definition of `beta_max` is likely appropriate: as-is, it is
defined somewhat arbitrarily to preserve a Fermi surface which stretches over
multiple k-points. This does not seem to be an important property in general.
Refinement of `sigma_min` is less clear; a too-small value of `sigma`
leads to a noisy variation of results with chemical potential.

Relative difference in regression test output observables is < ~ 10^{-5}.
Release-mode plots have non-visible difference.

Treatment of the delta function is generalized. Here we include an additional
delta function in which energies are grouped into fixed bins and the delta
function is finite when a pair of energies are in the same bin. This approach
can yield dramatically faster execution than the Gaussian delta function but
leads to noise which is not straightforward to control.